Add files via upload
This is the authors' version of the first paper published in this area by the authors.  It was published in the journal Computers and Industrial Engineering in December of 2024.   Its reference is as follows:
Triantaphyllou, E. and Yanase, J., 2024. The use of pairwise comparisons for decision making may lead to grossly inaccurate results. Computers & Industrial Engineering, Vol. 198, December 2024, p.110653.